Fix clang signedness errors in row/field numbers.
I just made the row/field numbers in the row and field classes signed.
This fixes an underflow in reverse iterators, but it creates some type
mismatches with places where we do use unsigned types.

+++ include/pqxx/row.hxx
@@ -110,7 +110,8 @@ public:
 
   void swap(row &) noexcept;						//[t11]
 
-  size_t rownumber() const noexcept { return m_index; }			//[t11]
+  /// Row number, assuming this is a real row and not end()/rend().
+  size_t rownumber() const noexcept { return size_t(m_index); }		//[t11]
 
   /**
    * @name Column information
